reworked ring buffer and added in dtaidtw. dtaidtw has flaws, need to fix

+++ Core/Inc/MPU_6050_Utils.h
@@ -128,7 +128,7 @@
 	((MPU_6050_ZG_EN) ? MPU_6050_ZG_FIFO_EN_MASK : 0) | \
 	((MPU_6050_ACCEL_EN) ? MPU_6050_ACCEL_FIFO_EN_MASK : 0)
 
-static const u8 accel_dplf_to_bandwidth_lut[] = {
+static const u16 accel_dplf_to_bandwidth_lut[] = {
 	260,
 	184,
 	94,
@@ -138,7 +138,7 @@ static const u8 accel_dplf_to_bandwidth_lut[] = {
 	5
 };
 
-static const u8 gyro_dplf_to_bandwidth_lut[] = {
+static const u16 gyro_dplf_to_bandwidth_lut[] = {
 	256,
 	188,
 	98,
